fix: resolve CI failures — Go example module, TS zstd compression, Rust clippy
- Move Go example to its own module (examples/go/resources_and_prompts/)
  so go vet ./... doesn't try to resolve the Go SDK submodule
- Add zstd-codec dependency to TypeScript SDK for proper zstd compression
  in raw sideband transfers (replaces missing fzstd)
- Add zstandard dependency to Python SDK pyproject.toml
- Fix Rust clippy type_complexity: extract ResourceHandler and
  CompletionKey/CompletionMap type aliases
- Fix Rust clippy io_other_error: use std::io::Error::other()
- Update CI workflow: scope go vet/test to cmd/internal/test/tests
  (excludes examples with separate modules)
- Fix e2e job to build pmcp binary to bin/pmcp (matches test harness path)
