+++ .github/workflows/ci.yml
@@ -14,8 +14,8 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version: '1.25'
-      - run: go vet ./...
-      - run: go test ./...
+      - run: go vet ./cmd/... ./internal/... ./test/... ./tests/...
+      - run: go test ./cmd/... ./internal/... ./test/... ./tests/...
       - run: go build ./cmd/protomcp/
 
   go-sdk:
@@ -72,7 +72,7 @@ jobs:
       - uses: actions/setup-node@v4
         with:
           node-version: '20'
-      - run: go build -o pmcp ./cmd/protomcp/
+      - run: go build -o bin/pmcp ./cmd/protomcp/
       - run: cd sdk/python && pip install -e .
       - run: cd sdk/typescript && npm ci && npm run build
-      - run: go test ./test/e2e/ -v
+      - run: go test ./test/e2e/ -v -timeout 60s
