+++ examples/go/resources_and_prompts/.gitignore
@@ -1,0 +1,1 @@
+resources-and-prompts-example
