+++ sdk/typescript/package.json
@@ -13,11 +13,12 @@
   "dependencies": {
     "protobufjs": "^7.0.0",
     "zod": "^3.22.0",
-    "zod-to-json-schema": "^3.22.0"
+    "zod-to-json-schema": "^3.22.0",
+    "zstd-codec": "^0.1.5"
   },
   "devDependencies": {
+    "@types/node": "^20.0.0",
     "typescript": "^5.3.0",
-    "vitest": "^1.0.0",
-    "@types/node": "^20.0.0"
+    "vitest": "^1.0.0"
   }
 }
