+++ sdk/typescript/src/zstd-codec.d.ts
@@ -1,0 +1,14 @@
+declare module 'zstd-codec' {
+  export class ZstdCodec {
+    static run(callback: (zstd: ZstdModule) => void): void;
+  }
+
+  interface ZstdModule {
+    Simple: new () => ZstdSimple;
+  }
+
+  interface ZstdSimple {
+    compress(data: Uint8Array, compressionLevel?: number): Uint8Array;
+    decompress(data: Uint8Array): Uint8Array;
+  }
+}
